backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5
Flask-JWT-Extended==4.5.3
scapy==2.5.0
python-dotenv==1.0.0
werkzeug==3.0.1
python-engineio==4.8.0
python-socketio==5.10.0
